#9 create generic existsInColByID method
